WIP: Move witness to primitives - broken
